src/main/java/com/leetcode/linkedlists/medium/SwapNodesInPairs_24.java
package com.leetcode.linkedlists.medium;

// Given a linked list, swap every two adjacent nodes and return its head.
//
// You may not modify the values in the list's nodes. Only nodes itself may be c
// hanged.
//
//
// Example 1:
//
//
// Input: head = [1,2,3,4]
// Output: [2,1,4,3]
//
//
// Example 2:
//
//
// Input: head = []
// Output: []
//
//
// Example 3:
//
//
// Input: head = [1]
// Output: [1]
//
//
//
// Constraints:
//
//
// The number of nodes in the list is in the range [0, 100].
// 0 <= Node.val <= 100
//
// Related Topics Linked List
// 👍 2813 👎 185

// leetcode submit region begin(Prohibit modification and deletion)

import com.leetcode.utils.LinkedList.ListNode;

/**
 * Definition for singly-linked list. public class ListNode { int val; ListNode next; ListNode() {}
 * ListNode(int val) { this.val = val; } ListNode(int val, ListNode next) { this.val = val;
 * this.next = next; } }
 */
/*
 O(n) Runtime: 0 ms, faster than 100.00% of Java online submissions for Swap Nodes in Pairs.
 O(1) Memory Usage: 36.6 MB, less than 8.51% of Java online submissions for Swap Nodes in Pairs.
*/
public class SwapNodesInPairs_24 {
  public ListNode swapPairs(ListNode head) {
    if (head == null || head.next == null) {
      return head;
    }
    ListNode next = head.next;
    head.next = swapPairs(head.next.next);
    next.next = head;
    return next;
  }
}
// leetcode submit region end(Prohibit modification and deletion)
